[rs-dam] Fix potential NPE.
Former-commit-id: 8c666ded6b9f0b1436a087542b71e97982326c1e

+++ rs-dam/dam/dam-service/src/main/java/fr/cnes/regards/modules/dam/service/entities/AbstractEntityService.java
@@ -626,7 +626,8 @@ public abstract class AbstractEntityService<F extends EntityFeature, U extends A
         datasets.remove(toDelete);
         // Remove relate files
         for (Map.Entry<DataType, DataFile> entry : toDelete.getFiles().entries()) {
-            if (localStorageService.isFileLocallyStored(toDelete, entry.getValue())) {
+            if ((entry != null) && (entry.getValue() != null)
+                    && localStorageService.isFileLocallyStored(toDelete, entry.getValue())) {
                 localStorageService.removeFile(toDelete, entry.getValue());
             }
         }
